refactor(examples): update apps to use new entry points
