+++ examples/next/src/app/page.tsx
@@ -1,23 +1,42 @@
 import {
-  Box,
   Circle,
   Flex,
   Grid,
   Hide,
   Show,
-  Skeleton,
   Square,
   Text,
 } from "@animareflection/ui";
-import Link from "next/link";
 
 import {
+  AccordionDemo,
   BadgeDemo,
+  BannerDemo,
+  BreadcrumbDemo,
+  ButtonDemo,
   CardDemo,
+  CarouselDemo,
+  CheckboxDemo,
+  CollapseDemo,
+  DrawerDemo,
+  FlyoutDemo,
+  IconDemo,
   ImageDemo,
+  InputDemo,
   KbdDemo,
+  MenuDemo,
+  ModalDemo,
+  NumberInputDemo,
+  RadioGroupDemo,
+  SkeletonToggleDemo,
+  SliderDemo,
   SpinnerDemo,
-} from "components/core";
+  TabsDemo,
+  TextareaDemo,
+  ToastDemo,
+  ToggleDemo,
+  TooltipDemo,
+} from "components";
 
 const HomePage = () => (
   <Flex direction="column" h="100%" align="center" gap={4} pt={12}>
@@ -60,11 +79,34 @@ const HomePage = () => (
     </Grid>
 
     <Flex direction="column" w="100%" gap={2} p={4}>
+      <ButtonDemo />
+      <AccordionDemo />
+      <DrawerDemo />
+      <IconDemo />
+      <MenuDemo />
+      <ModalDemo />
+      <SliderDemo />
       <BadgeDemo />
       <CardDemo />
       <ImageDemo />
       <SpinnerDemo />
+      <TabsDemo />
+      <ToggleDemo />
+      <SkeletonToggleDemo />
+      <CheckboxDemo />
       <KbdDemo />
+      <CollapseDemo />
+      <BannerDemo />
+      <InputDemo />
+      <NumberInputDemo />
+      <TextareaDemo />
+      <FlyoutDemo />
+      <TooltipDemo />
+      <ToastDemo />
+      <CarouselDemo />
+      <BreadcrumbDemo />
+      <RadioGroupDemo />
+
       <Hide below="md">
         <Text color="brand.primary.500" fontSize="3xl" fontWeight="bold">
           Hide below md breakpoint
@@ -85,12 +127,7 @@ const HomePage = () => (
           Show from md breakpoint and up
         </Text>
       </Show>
-      <Skeleton variant="circle" w={12} h={12} />
     </Flex>
-
-    <Box color="blue" _hover={{ textDecoration: "underline" }}>
-      <Link href="/client">Go to client page demo</Link>
-    </Box>
   </Flex>
 );
