+++ examples/next/src/components/CarouselDemo.tsx
@@ -1,7 +1,7 @@
-import { Carousel } from "@animareflection/ui/client";
+import { Carousel } from "@animareflection/ui";
 import { Image } from "@animareflection/ui/next";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const IMAGE_SOURCES = [
   "https://tinyurl.com/5b6ka8jd",
