+++ examples/next/src/components/CheckboxDemo.tsx
@@ -1,9 +1,7 @@
-"use client";
-
-import { Checkbox } from "@animareflection/ui/client";
+import { Checkbox } from "@animareflection/ui";
 import { useState } from "react";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const CheckboxDemo = () => {
   const [isChecked, setIsChecked] = useState(true);
