+++ examples/next/src/components/InputDemo.tsx
@@ -1,6 +1,6 @@
-import { Input } from "@animareflection/ui/client";
+import { Input } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const InputDemo = () => (
   <Wrapper title="Input">
