+++ examples/next/src/components/SkeletonToggleDemo.tsx
@@ -2,10 +2,9 @@
 
 import { useState } from "react";
 
-import { Skeleton, Text } from "@animareflection/ui";
-import { Button } from "@animareflection/ui/client";
+import { Button, Skeleton, Text } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const SkeletonToggleDemo = () => {
   const [isLoaded, setIsLoaded] = useState(false);
