+++ examples/next/src/components/SpinnerDemo.tsx
@@ -1,6 +1,6 @@
 import { Spinner } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const SpinnerDemo = () => (
   <Wrapper title="Spinner">
