+++ examples/vite-react/src/components/CollapseDemo.tsx
@@ -1,7 +1,6 @@
-import { Flex, Image, Text } from "@animareflection/ui";
-import { Collapse } from "@animareflection/ui/client";
+import { Collapse, Flex, Image, Text } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const CollapseDemo = () => (
   <Wrapper title="Collapse">
