+++ examples/vite-react/src/components/ImageDemo.tsx
@@ -1,6 +1,6 @@
 import { Image } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const ImageDemo = () => (
   <Wrapper title="Image">
