+++ examples/vite-react/src/components/SkeletonToggleDemo.tsx
@@ -1,11 +1,7 @@
-"use client";
-
+import { Button, Skeleton, Text } from "@animareflection/ui";
 import { useState } from "react";
 
-import { Skeleton, Text } from "@animareflection/ui";
-import { Button } from "@animareflection/ui/client";
-
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const SkeletonToggleDemo = () => {
   const [isLoaded, setIsLoaded] = useState(false);
