+++ examples/vite-react/src/components/ToastDemo.tsx
@@ -1,7 +1,6 @@
-import { toast } from "@animareflection/ui";
-import { Button, Toast, Toaster } from "@animareflection/ui/client";
+import { Button, toast, Toast, Toaster } from "@animareflection/ui";
 
-import { Wrapper } from "components/core";
+import { Wrapper } from "components";
 
 const ToastDemo = () => {
   const promise = () => {
